Add example data
A Python-based tensile test data mapping script (Jupyter Notebook) is included in our repository as an example. The input data is included in the folder "tensile_test_data" as well. The output of this mapping script is provided in two formats as "S355_data_tto.ttl" and "S355_data_tto.rdf".
The latter may be loaded in a triple store.